RrdGraph.js: gprint: fix check for undefined format in case of vdef
The GPRINT of a VDEF doesn't have a CF. For example:
> "GPRINT:avg:LAST:%5.1lf%s Last"
> "GPRINT:tot:%5.1lf%s Total\l"

The format will be passed via the cf variable, so format is not defined. But it
didn't show up in the legend. Firebug showed variable format isn't undefined,
but an empty string in case of VDEF. So just check for falsy.