Remove unused getStringAt, move implementation
Rename switch_endian to littleEndian to be more explicit (the meaning
is now inversed too). Make getEndianByteAt private since it is an
implementation detail.

Drop getStringAt as this is not used in this library. Move getCharAt
above getCStringAt since they are more related.

Move the methods definition for the current implementation to a new
function.